+++ src/main/java/io/github/abhinavminhas/ShadowRootAssist.java
@@ -1,20 +1,18 @@
 package io.github.abhinavminhas;
 
-import java.util.Map;
+import java.time.Duration;
 import org.openqa.selenium.JavascriptExecutor;
+import org.openqa.selenium.SearchContext;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebDriverException;
 import org.openqa.selenium.support.ui.ExpectedCondition;
 import org.openqa.selenium.support.ui.WebDriverWait;
-import org.openqa.selenium.WebElement;
-import org.openqa.selenium.remote.RemoteWebDriver;
-import org.openqa.selenium.remote.RemoteWebElement;
 
 /** 
  * 'ShadowRootAssist' class to support digging of shadow roots in DOM.
  * 
  * @author 	Abhinav Minhas
- * @version 1.0.0
+ * @version 2.0.0
  * @since 	01-01-2022
  */
 public class ShadowRootAssist {
@@ -26,26 +24,18 @@ public class ShadowRootAssist {
 	 * @param shadowRootSelector 			Shadow root element selector (probably jQuery or CssSelectors).
 	 * @param timeInSeconds 				Wait time in seconds.
 	 * @param pollingIntervalInMilliseconds	Polling interval time in milliseconds.
-	 * @return								Shadow root web element.
+	 * @return								Shadow root (SearchContext).
 	 * @exception WebDriverException		Throws - 'WebDriverException' in case any shadow root element is not found.
 	 */
-	public static WebElement getShadowRootElement(WebDriver webDriver, String shadowRootSelector, int timeInSeconds, int pollingIntervalInMilliseconds) {
-		WebElement requiredShadowRoot = null;
+	public static SearchContext getShadowRootElement(WebDriver webDriver, String shadowRootSelector, int timeInSeconds, int pollingIntervalInMilliseconds) {
+		SearchContext requiredShadowRoot = null;
 		String shadowRootQuerySelector = "return document.querySelector('%s').shadowRoot";
  		String shadowRootElement = String.format(shadowRootQuerySelector, shadowRootSelector);
 		try {
-			WebDriverWait webDriverWait = new WebDriverWait(webDriver, timeInSeconds, pollingIntervalInMilliseconds);
+			WebDriverWait webDriverWait = new WebDriverWait(webDriver, Duration.ofSeconds(timeInSeconds), Duration.ofMillis(pollingIntervalInMilliseconds));
 			webDriverWait.until((ExpectedCondition<Boolean>) wd -> ((JavascriptExecutor)webDriver).executeScript(shadowRootElement) != null);
 			Object returnedObject = ((JavascriptExecutor)webDriver).executeScript(shadowRootElement);
-			if (returnedObject instanceof Map) {
-				@SuppressWarnings("unchecked")
-				Map<String, Object> map = (Map<String, Object>)returnedObject;
-				RemoteWebElement remoteWebElement = new RemoteWebElement();
-				remoteWebElement.setParent((RemoteWebDriver)webDriver);
-				remoteWebElement.setId((String)map.values().iterator().next());
-				requiredShadowRoot = remoteWebElement;
-			} else
-				requiredShadowRoot = (WebElement)returnedObject;
+			requiredShadowRoot = (SearchContext)returnedObject;
 		} catch(WebDriverException ex) {
 			throw new WebDriverException(String.format("%s: Shadow root element for selector '%s' Not Found.", new Object(){}.getClass().getEnclosingMethod().getName(), shadowRootSelector));
 		}
@@ -59,11 +49,11 @@ public class ShadowRootAssist {
 	 * @param shadowRootSelectors			List of shadow root element selectors (probably jQuery or CssSelectors) separated by '&gt;'.
 	 * @param timeInSeconds					Wait time in seconds.
 	 * @param pollingIntervalInMilliseconds Polling interval time in milliseconds.
-	 * @return								Nested shadow root web element.
+	 * @return								Nested shadow root (SearchContext).
 	 * @exception WebDriverException		Throws - 'WebDriverException' in case any shadow root element is not found in the nested hierarchy.
 	 */
-	public static WebElement getNestedShadowRootElement(WebDriver webDriver, String shadowRootSelectors, int timeInSeconds, int pollingIntervalInMilliseconds) {
-		WebElement requiredShadowRoot = null;
+	public static SearchContext getNestedShadowRootElement(WebDriver webDriver, String shadowRootSelectors, int timeInSeconds, int pollingIntervalInMilliseconds) {
+		SearchContext requiredShadowRoot = null;
         String[] listShadowRootSelectors = shadowRootSelectors.split(">");
         for (int i = 0; i < listShadowRootSelectors.length ; i++) { listShadowRootSelectors[i] = listShadowRootSelectors[i].trim(); }
         String shadowRootQuerySelector = ".querySelector('%s').shadowRoot";
@@ -74,18 +64,10 @@ public class ShadowRootAssist {
             shadowRootQueryString += tempQueryString;
             String shadowRootElement = String.format(documentReturn, shadowRootQueryString);
             try {
-            	WebDriverWait webDriverWait = new WebDriverWait(webDriver, timeInSeconds, pollingIntervalInMilliseconds);
+            	WebDriverWait webDriverWait = new WebDriverWait(webDriver, Duration.ofSeconds(timeInSeconds), Duration.ofMillis(pollingIntervalInMilliseconds));
                 webDriverWait.until((ExpectedCondition<Boolean>) wd -> ((JavascriptExecutor)webDriver).executeScript(shadowRootElement) != null);
     			Object returnedObject = ((JavascriptExecutor)webDriver).executeScript(shadowRootElement);
-    			if (returnedObject instanceof Map) {
-    				@SuppressWarnings("unchecked")
-    				Map<String, Object> map = (Map<String, Object>)returnedObject;
-    				RemoteWebElement remoteWebElement = new RemoteWebElement();
-    				remoteWebElement.setParent((RemoteWebDriver)webDriver);
-    				remoteWebElement.setId((String)map.values().iterator().next());
-    				requiredShadowRoot = remoteWebElement;
-    			} else
-    				requiredShadowRoot = (WebElement)returnedObject;
+    			requiredShadowRoot = (SearchContext)returnedObject;
             } catch (WebDriverException ex) {
             	throw new WebDriverException(String.format("%s: Nested shadow root element for selector '%s' in DOM hierarchy '%s' Not Found.", new Object(){}.getClass().getEnclosingMethod().getName(), shadowRoot, shadowRootSelectors));
             }
@@ -109,22 +91,11 @@ public class ShadowRootAssist {
 		String shadowRootQuerySelector = "return document.querySelector('%s').shadowRoot";
 		String shadowRootElement = String.format(shadowRootQuerySelector, shadowRootSelector);
         try {
-        	WebDriverWait webDriverWait = new WebDriverWait(webDriver, timeInSeconds, pollingIntervalInMilliseconds);
+        	WebDriverWait webDriverWait = new WebDriverWait(webDriver, Duration.ofSeconds(timeInSeconds), Duration.ofMillis(pollingIntervalInMilliseconds));
 			webDriverWait.until((ExpectedCondition<Boolean>) wd -> ((JavascriptExecutor)webDriver).executeScript(shadowRootElement) != null); 
 			Object returnedObject = ((JavascriptExecutor)webDriver).executeScript(shadowRootElement);
-			if (returnedObject instanceof Map) {
-				@SuppressWarnings("unchecked")
-				Map<String, Object> map = (Map<String, Object>)returnedObject;
-				RemoteWebElement remoteWebElement = new RemoteWebElement();
-				remoteWebElement.setParent((RemoteWebDriver)webDriver);
-				remoteWebElement.setId((String)map.values().iterator().next());
-				Object requiredShadowRoot = remoteWebElement;
-				webDriverWait.until(item -> requiredShadowRoot instanceof WebElement);
-				isPresent = true;
-			} else {
-				webDriverWait.until(item -> returnedObject instanceof WebElement);
-				isPresent = true;
-			}
+			webDriverWait.until(item -> returnedObject instanceof SearchContext);
+			isPresent = true;
         } catch (WebDriverException ex) {
         	if (throwError)
                 throw new WebDriverException(String.format("%s: Shadow root element for selector '%s' Not Found.", new Object(){}.getClass().getEnclosingMethod().getName(), shadowRootSelector));
@@ -157,22 +128,11 @@ public class ShadowRootAssist {
             shadowRootQueryString += tempQueryString;
             String shadowRootElement = String.format(documentReturn, shadowRootQueryString);
             try {
-            	WebDriverWait webDriverWait = new WebDriverWait(webDriver, timeInSeconds, pollingIntervalInMilliseconds);
+            	WebDriverWait webDriverWait = new WebDriverWait(webDriver, Duration.ofSeconds(timeInSeconds), Duration.ofMillis(pollingIntervalInMilliseconds));
                 webDriverWait.until((ExpectedCondition<Boolean>) wd -> ((JavascriptExecutor)webDriver).executeScript(shadowRootElement) != null);
     			Object returnedObject = ((JavascriptExecutor)webDriver).executeScript(shadowRootElement);
-    			if (returnedObject instanceof Map) {
-    				@SuppressWarnings("unchecked")
-    				Map<String, Object> map = (Map<String, Object>)returnedObject;
-    				RemoteWebElement remoteWebElement = new RemoteWebElement();
-    				remoteWebElement.setParent((RemoteWebDriver)webDriver);
-    				remoteWebElement.setId((String)map.values().iterator().next());
-    				Object requiredShadowRoot = remoteWebElement;
-    				webDriverWait.until(item -> requiredShadowRoot instanceof WebElement);
-    				isPresent = true;
-    			} else {
-    				webDriverWait.until(item -> returnedObject instanceof WebElement);
-    				isPresent = true;
-    			}
+    			webDriverWait.until(item -> returnedObject instanceof SearchContext);
+				isPresent = true;
             } catch (WebDriverException ex) {
             	if (throwError)
             		throw new WebDriverException(String.format("%s: Nested shadow root element for selector '%s' in DOM hierarchy '%s' Not Found.", new Object(){}.getClass().getEnclosingMethod().getName(), shadowRoot, shadowRootSelectors));