+++ src/test/java/shadowrootdigger/tests/ShadowRootDiggerChromeSettingsTests.java
@@ -1,8 +1,10 @@
 package shadowrootdigger.tests;
 
+import java.time.Duration;
 import java.util.List;
 import org.openqa.selenium.By;
 import org.openqa.selenium.ElementNotInteractableException;
+import org.openqa.selenium.SearchContext;
 import org.openqa.selenium.TimeoutException;
 import org.openqa.selenium.WebDriverException;
 import org.openqa.selenium.WebElement;
@@ -31,7 +33,7 @@ public class ShadowRootDiggerChromeSettingsTests extends TestBase {
 	@Test(description = CHROME_SETTINGS_TESTS)
 	public void test_getShadowRootElement_ChromeSettings_ShadowRootElementExists() {
         webDriver.navigate().to("chrome://settings/clearBrowserData");
-        WebElement clearBrowsingTab = ShadowRootAssist.getShadowRootElement(webDriver, existsShadowRootElement, 20 , 2000);
+        SearchContext clearBrowsingTab = ShadowRootAssist.getShadowRootElement(webDriver, existsShadowRootElement, 20 , 2000);
         Assert.assertNotNull(clearBrowsingTab);
     }
 	
@@ -48,15 +50,43 @@ public class ShadowRootDiggerChromeSettingsTests extends TestBase {
     }
 	
 	@Test(description = CHROME_SETTINGS_TESTS)
+	public void test_getShadowRootElement_ChromeSettings_ImplicitWaitManipulationCheck_Exists() {
+		webDriver.manage().timeouts().implicitlyWait(Duration.ofSeconds(0));
+        long implicitWaitBefore = webDriver.manage().timeouts().getImplicitWaitTimeout().toMillis();
+        Assert.assertEquals(implicitWaitBefore, 0);
+        webDriver.navigate().to("chrome://settings/clearBrowserData");
+        ShadowRootAssist.getShadowRootElement(webDriver, existsShadowRootElement, 20 , 2000);
+        long implicitWaitAfter = webDriver.manage().timeouts().getImplicitWaitTimeout().toMillis();
+        Assert.assertEquals(implicitWaitBefore, implicitWaitAfter);
+	}
+	
+	@Test(description = CHROME_SETTINGS_TESTS)
+	public void test_getShadowRootElement_ChromeSettings_ImplicitWaitManipulationCheck_NotExists() {
+		webDriver.manage().timeouts().implicitlyWait(Duration.ofSeconds(0));
+        long implicitWaitBefore = webDriver.manage().timeouts().getImplicitWaitTimeout().toMillis();
+        Assert.assertEquals(implicitWaitBefore, 0);
+        webDriver.navigate().to("chrome://settings/clearBrowserData");
+        try {
+            ShadowRootAssist.getShadowRootElement(webDriver, notExistsShadowRootElement, 20 , 2000);
+            Assert.fail("No Exception Thrown.");
+        }
+        catch (AssertionError ex) { throw ex; }
+        catch (WebDriverException ex) {
+        	long implicitWaitAfter = webDriver.manage().timeouts().getImplicitWaitTimeout().toMillis();
+            Assert.assertEquals(implicitWaitBefore, implicitWaitAfter);
+        }
+    }
+	
+	@Test(description = CHROME_SETTINGS_TESTS)
 	public void test_getNestedShadowRootElement_ChromeSettings_ClearChromeData() {
         webDriver.navigate().to("chrome://settings/clearBrowserData");
-        WebElement clearBrowsingTab = ShadowRootAssist.getNestedShadowRootElement(webDriver, tabRootElement, 20 , 2000);
+        SearchContext clearBrowsingTab = ShadowRootAssist.getNestedShadowRootElement(webDriver, tabRootElement, 20 , 2000);
         int count = 0;
         do {
         	count++;
         	try {
         		WebElement requiredElement = clearBrowsingTab.findElements(By.cssSelector(divTabIdentifier)).iterator().next();
-        		WebDriverWait webDriverWait = new WebDriverWait(webDriver, 10, 1000);
+        		WebDriverWait webDriverWait = new WebDriverWait(webDriver, Duration.ofSeconds(10) , Duration.ofMillis(1000));
         		webDriverWait.until((ExpectedCondition<Boolean>) wd -> requiredElement.getText().equals("Basic") && requiredElement.isDisplayed() && requiredElement.isEnabled());
         		requiredElement.click();
         		break;
@@ -68,10 +98,10 @@ public class ShadowRootDiggerChromeSettingsTests extends TestBase {
         		/* Retry */
         	}
         } while (count <= 1);
-        WebElement settingsDropdownMenu = ShadowRootAssist.getNestedShadowRootElement(webDriver, settingsDropdownMenuRootElement, 20 , 2000);
+        SearchContext settingsDropdownMenu = ShadowRootAssist.getNestedShadowRootElement(webDriver, settingsDropdownMenuRootElement, 20 , 2000);
         WebElement timeRangeSelect = settingsDropdownMenu.findElement(By.cssSelector(selectTimeRangeIdentifier));
         new Select(timeRangeSelect).selectByVisibleText("Last hour");
-        WebElement clearBrowsingDataDialog = ShadowRootAssist.getNestedShadowRootElement(webDriver, clearBrowsingDataDialogRootElement, 20 , 2000);
+        SearchContext clearBrowsingDataDialog = ShadowRootAssist.getNestedShadowRootElement(webDriver, clearBrowsingDataDialogRootElement, 20 , 2000);
         List<WebElement> basicCheckboxes = clearBrowsingDataDialog.findElements(By.cssSelector(basicTabCheckboxesIdentifier));
         for (WebElement checkbox : basicCheckboxes) {
         	String isCheckboxChecked = checkbox.getAttribute("checked");
@@ -94,6 +124,34 @@ public class ShadowRootDiggerChromeSettingsTests extends TestBase {
 	}
 	
 	@Test(description = CHROME_SETTINGS_TESTS)
+	public void test_getNestedShadowRootElement_ChromeSettings_ImplicitWaitManipulationCheck_Exists() {
+		webDriver.manage().timeouts().implicitlyWait(Duration.ofSeconds(0));
+        long implicitWaitBefore = webDriver.manage().timeouts().getImplicitWaitTimeout().toMillis();
+        Assert.assertEquals(implicitWaitBefore, 0);
+        webDriver.navigate().to("chrome://settings/clearBrowserData");
+        ShadowRootAssist.getNestedShadowRootElement(webDriver, tabRootElement, 20 , 2000);
+        long implicitWaitAfter = webDriver.manage().timeouts().getImplicitWaitTimeout().toMillis();
+        Assert.assertEquals(implicitWaitBefore, implicitWaitAfter);
+    }
+	
+	@Test(description = CHROME_SETTINGS_TESTS)
+	public void test_getNestedShadowRootElement_ChromeSettings_ImplicitWaitManipulationCheck_NotExists() {
+		webDriver.manage().timeouts().implicitlyWait(Duration.ofSeconds(0));
+		long implicitWaitBefore = webDriver.manage().timeouts().getImplicitWaitTimeout().toMillis();
+        Assert.assertEquals(implicitWaitBefore, 0);
+        webDriver.navigate().to("chrome://settings/clearBrowserData");
+        try {
+            ShadowRootAssist.getNestedShadowRootElement(webDriver, notExistsNestedShadowRootElement, 20 , 2000);
+            Assert.fail("No Exception Thrown.");
+        }
+        catch (AssertionError ex) { throw ex; }
+        catch (WebDriverException ex) {
+        	long implicitWaitAfter = webDriver.manage().timeouts().getImplicitWaitTimeout().toMillis();
+            Assert.assertEquals(implicitWaitBefore, implicitWaitAfter);
+        }
+    }
+	
+	@Test(description = CHROME_SETTINGS_TESTS)
 	public void test_isShadowRootElementPresent_ChromeSettings_ShadowRootExists() {
 		webDriver.navigate().to("chrome://settings/clearBrowserData");
 		Boolean exists = ShadowRootAssist.isShadowRootElementPresent(webDriver, existsShadowRootElement, false, 20, 2000);
@@ -113,6 +171,37 @@ public class ShadowRootDiggerChromeSettingsTests extends TestBase {
         catch (AssertionError ex)  { throw ex; }
         catch (WebDriverException ex) { Assert.assertTrue(ex.getMessage().contains(expectedErrorMessage)); }
 	}
+	
+	@Test(description = CHROME_SETTINGS_TESTS)
+	public void test_isShadowRootElementPresent_ChromeSettings_ImplicitWaitManipulationCheck_Exists() {
+		webDriver.manage().timeouts().implicitlyWait(Duration.ofSeconds(0));
+        long implicitWaitBefore = webDriver.manage().timeouts().getImplicitWaitTimeout().toMillis();
+        Assert.assertEquals(implicitWaitBefore, 0);
+        webDriver.navigate().to("chrome://settings/clearBrowserData");
+        ShadowRootAssist.isShadowRootElementPresent(webDriver, existsShadowRootElement, false, 20 , 2000);
+        long implicitWaitAfter = webDriver.manage().timeouts().getImplicitWaitTimeout().toMillis();
+        Assert.assertEquals(implicitWaitBefore, implicitWaitAfter);
+    }
+	
+	@Test(description = CHROME_SETTINGS_TESTS)
+	public void test_isShadowRootElementPresent_ChromeSettings_ImplicitWaitManipulationCheck_NotExists() {
+		webDriver.manage().timeouts().implicitlyWait(Duration.ofSeconds(0));
+		long implicitWaitBefore = webDriver.manage().timeouts().getImplicitWaitTimeout().toMillis();
+		Assert.assertEquals(implicitWaitBefore, 0);
+        webDriver.navigate().to("chrome://settings/clearBrowserData");
+        ShadowRootAssist.isShadowRootElementPresent(webDriver, notExistsShadowRootElement, false, 20 , 2000);
+        long implicitWaitAfter = webDriver.manage().timeouts().getImplicitWaitTimeout().toMillis();
+        Assert.assertEquals(implicitWaitBefore, implicitWaitAfter);
+        try {
+            ShadowRootAssist.isShadowRootElementPresent(webDriver, notExistsShadowRootElement, true, 20 , 2000);
+            Assert.fail("No Exception Thrown.");
+        }
+        catch (AssertionError ex) { throw ex; }
+        catch (WebDriverException ex) {
+        	implicitWaitAfter = webDriver.manage().timeouts().getImplicitWaitTimeout().toMillis();
+            Assert.assertEquals(implicitWaitBefore, implicitWaitAfter);
+        }
+    }
 	
 	@Test(description = CHROME_SETTINGS_TESTS)
 	public void test_isNestedShadowRootElementPresent_ChromeSettings_NestedShadowRootExists() {
@@ -139,4 +228,34 @@ public class ShadowRootDiggerChromeSettingsTests extends TestBase {
         catch (WebDriverException ex) { Assert.assertTrue(ex.getMessage().contains(expectedErrorMessage)); }
 	}
 	
+	@Test(description = CHROME_SETTINGS_TESTS)
+	public void test_isNestedShadowRootElementPresent_ChromeSettings_ImplicitWaitManipulationCheck_Exists() {
+		webDriver.manage().timeouts().implicitlyWait(Duration.ofSeconds(0));
+        long implicitWaitBefore = webDriver.manage().timeouts().getImplicitWaitTimeout().toMillis();
+        Assert.assertEquals(implicitWaitBefore, 0);
+        webDriver.navigate().to("chrome://settings/clearBrowserData");
+        ShadowRootAssist.isNestedShadowRootElementPresent(webDriver, tabRootElement, false, 20 , 2000);
+        long implicitWaitAfter = webDriver.manage().timeouts().getImplicitWaitTimeout().toMillis();
+        Assert.assertEquals(implicitWaitBefore, implicitWaitAfter);
+    }
+	
+	@Test(description = CHROME_SETTINGS_TESTS)
+	public void test_isNestedShadowRootElementPresent_ChromeSettings_ImplicitWaitManipulationCheck_NotExists() {
+		webDriver.manage().timeouts().implicitlyWait(Duration.ofSeconds(0));
+        long implicitWaitBefore = webDriver.manage().timeouts().getImplicitWaitTimeout().toMillis();
+        Assert.assertEquals(implicitWaitBefore, 0);
+        webDriver.navigate().to("chrome://settings/clearBrowserData");
+        ShadowRootAssist.isNestedShadowRootElementPresent(webDriver, notExistsNestedShadowRootElement, false, 20 , 2000);
+        long implicitWaitAfter = webDriver.manage().timeouts().getImplicitWaitTimeout().toMillis();
+        Assert.assertEquals(implicitWaitBefore, implicitWaitAfter);
+        try {
+            ShadowRootAssist.isNestedShadowRootElementPresent(webDriver, notExistsNestedShadowRootElement, true, 20 , 2000);
+            Assert.fail("No Exception Thrown.");
+        }
+        catch (AssertionError ex) { throw ex; }
+        catch (WebDriverException ex) {
+        	implicitWaitAfter = webDriver.manage().timeouts().getImplicitWaitTimeout().toMillis();
+            Assert.assertEquals(implicitWaitBefore, implicitWaitAfter);
+        }
+    }
 }